Move regex machinery into ERTS
