+++ assets/js/hologram.mjs
@@ -20,7 +20,6 @@ import JsInterop from "./js_interop.mjs";
 import MemoryStorage from "./memory_storage.mjs";
 import Operation from "./operation.mjs";
 import PerformanceTimer from "./performance_timer.mjs";
-import RegexEngine from "./regex/regex_engine.mjs";
 import Renderer from "./renderer.mjs";
 import Serializer from "./serializer.mjs";
 import Sse from "./sse.mjs";
@@ -76,7 +75,6 @@ export default class Hologram {
     HologramInterpreterError: HologramInterpreterError,
     Interpreter: Interpreter,
     MemoryStorage: MemoryStorage,
-    RegexEngine: RegexEngine,
     Type: Type,
     Utils: Utils,
   };
