+++ test/javascript/erts/regex/regex_analyzer_test.mjs
@@ -3,10 +3,10 @@
 import {
   assert,
   defineGlobalErlangAndElixirModules,
-} from "../support/helpers.mjs";
+} from "../../support/helpers.mjs";
 
-import RegexAnalyzer from "../../../assets/js/regex/regex_analyzer.mjs";
-import RegexParser from "../../../assets/js/regex/regex_parser.mjs";
+import RegexAnalyzer from "../../../../assets/js/erts/regex/regex_analyzer.mjs";
+import RegexParser from "../../../../assets/js/erts/regex/regex_parser.mjs";
 
 defineGlobalErlangAndElixirModules();
