+++ test/javascript/support/helpers.mjs
@@ -27,8 +27,8 @@ import Erlang_Uri_String from "../../../assets/js/erlang/uri_string.mjs";
 import HologramBoxedError from "../../../assets/js/errors/boxed_error.mjs";
 import HologramInterpreterError from "../../../assets/js/errors/interpreter_error.mjs";
 import Interpreter from "../../../assets/js/interpreter.mjs";
-import RegexParseError from "../../../assets/js/regex/regex_parse_error.mjs";
-import RegexParser from "../../../assets/js/regex/regex_parser.mjs";
+import RegexParseError from "../../../assets/js/erts/regex/regex_parse_error.mjs";
+import RegexParser from "../../../assets/js/erts/regex/regex_parser.mjs";
 import Renderer from "../../../assets/js/renderer.mjs";
 import Serializer from "../../../assets/js/serializer.mjs";
 import Type from "../../../assets/js/type.mjs";
